services/daily/requirements.txt
Flask==1.0.3
Flask-Caching==1.7.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
nba-sdk==0.1.0
psycopg2-binary==2.8.3
redis==3.2.1
